Extract MaybeCompressedFileReader
